ttools: fix upload matcher to cope with empty input tables
The CDS upload service currently fails with an error, at least under
some circumstances (selection=all) if there are no rows in the input
table.  So add logic which is able, configurably, to fail the match
with a well-behaved IOException in the case of zero input rows rather
than attempt the match.

Once the CDS service copes with empty upload tables this should
perhaps be switched back, since an empty output table may be of
some use (at least you can get the columns).  Or maybe not, you
probably don't want to keep the result.