ttools: add HEALPix presorting to CDS upload skymatch
Sorting the rows of an uploaded table by HEALPix pixel index for a
multi-block upload match job should in theory improve performance.
Add a parameter presort to the cdsskymatch task to turn this on.
There are some disadvantages though, so set it false by default.

It also makes it harder to do progress monitoring, so for now do
not provide this option in topcat.  For the topact xmatch progress
bar to work with this option would require (I think) a new optional
method getRowCount on ConeQuerySequenceFactory.