topcat: fix TAP load window to stay open by default
All table load dialogues by default are disposed when a table is loaded,
but you can change this behaviour with the "Stay Open" option.
For TAP it's usually pretty annoying for it to keep closing, so set
it pinned open by default.  Maybe I should do this for some other
load dialogues too?