ttools: generalise SkyCoord handling of null radius
Provide API configuration options for what happens when volume-type
SkyCoord user values have a blank radius - you might want to ignore
the point or assume it's on the unit sphere.

This was originally done so that the Sphere plot could default to
plotting points on the unit sphere rather than plotting no points
when the radius part is blank, to give something more like the
behaviour of the v3 sky/sphere plot.  However, it doesn't really
work because of other parts of the coordinate specification
(Coord.isRequired is a scalar, not per user-info), and in any case
it's not a good idea from a GUI point of view because you want to
populate the radius coordinate with some value by way of example.
So in this commit there is no change of behaviour, but it provides
more flexilibity in case it's useful in the future.