gbin: add user-level documentation
The GBIN format is now documented in the STILTS and TOPCAT user
documents.

The documentation text is stored in the GBIN package
and imported into SUN/253; the same could be done for SUN/256,
but at present the other formats are just documented in a couple
of lines there, so follow the same practice.  That was done
(presumably?) to avoid repeated text, but importing the text
from an external package like I have done here would make it
work.  Maybe fix it for the other formats one day.