topcat: fix regression bug (auth failure) in JDBC load dialogue
When I added the TAP load dialogue to the ControlWindow toolbar
just before release of TOPCAT v4.2 (4289ffb4), I inadvertently
introduced a bug which meant that the SQL load dialogue doesn't
work any more - authentication fails.  I haven't tracked down
exactly why this happens, but it's related to the fact that
the LoadWindow was being instantiated during ControlWindow
construction (before that change it was not done until the user
hit the Load button).

Modified ControlWindow so that LoadWindow is not constructed
until it's needed, which fixes that bug.  It's possible there
were other unnoticed negative consequences of the earlier change;
this update may fix those too.

Both the problem and the fix are fairly inelegant, because application
startup in topcat (implemented in the very old classes Driver,
ControlWindow, LoadWindow) is messy and badly designed.
Added a note in ControlWindow to this effect.