ttools: add PlotterParameter
The layerN parameter is now provided by a custom parameter
type PlotterParameter.  The values of this parameter no longer
map 1:1 to Plotter instances, since it previously had far too many
possible values
(mark-auto, mark-flat, mark-aux, ... link-auto, link-flat, link-aux, ....),
which made it unwieldy to (auto-)document.

Now, for the special case of ShapeModePlotters, this parameter
just specifies the ShapeForm, and an auxiliary (shading) parameter
is used to specify ShapeMode.  For other plotter types, this parameter
specifies the Plotter in full.