ttools: parameter documentation now includes value class
The auto-generated parameter documentation (in SUN/256 and on the
console using the help parameter) now include information about
what class the parameter represents.  This is useful/necessary
for programmatic use of the parameter system, since you may want
to submit typed values to a MapEnvironment as an alternative to
string values, so you need to know what type is permitted.

However, this is a bit problematic, since for the first time the
TTOOLS API is being exposed in the STILTS UI.  Various issues
with that, including the fact that the javadocs are not currently
shipped as part of the STILTS package, which they really should be
if some of the classes are relevant.  In the long term, I'd like
to split the plot2 classes out of STILTS altogether and expose
only that API, but in the meantime, reference the javadocs
published from my web page as a stopgap.

I'm also not quite sure about the aesthetic presentation of the
parameter types in the Usage sections of the auto-generated task
documentation.  Maybe revisit the details at some point.