ttools: partially address issues with swing plot2task animation
Swing animation is not parallelised.  This meant that for slow plots,
most frames are never seen with swing animations from Plot2Tasks.
Fix this by replacing SwingUtilities.invokeLater with invokeAndWait.
However the real fix would be to parallelise the plotting.
It wouldn't be too hard, but I'm not doing it now.
Add comments to the source code to describe what would have to be
done.