ttools: rework plot2 data coordinate metadata
Replace the use of ValueInfos with a new class Input to describe
the user-facing data coordinates used in the plot2 framework.
Input has a related class InputMeta which contains all the
user-directed metadata (various names and auto-documentation text items).
This mirrors the use of the Config and ConfigMeta classes for
scalar plot2 configuration items.

This allows much richer and more plot2-appropriate metadata to be
configured by the plotting components themeselves, so that console
on-line help and auto-generated documentation can be more comprehensive.
It also cleans up (a bit) the rather horrible and non-obvious
distinction between Coord and UserCoord in the API.
UserCoord is now gone, replaced by Input, which is less confusing.

It also further decouples the plot2 framework from STIL (good),
by removing use of ValueInfos from a lot of places, though they
still feature in DataSpec.