ttools: replace PlotterParameter with LayerTypeParameter
The Plotter is not the right abstraction for the user interface,
introduce a new class LayerType (a slight abstraction from Plotter)
which is, and give it a dedicated parameter, rather than fudging
the PlotterParameter as before.

Probably the LayerType should be used all the way back to PlotType
definitions, it would clean up code in topcat too.  But not today.