ttools: add XML description strings for LayerTypes
As currently implemented this requires a getPlotterDescription
method added to Plotter.  Really, that's the wrong level for this
to go at, the description should just be attached to the LayerType.
For now, LayerType only exists in the task invocation layer, but
I should at some point put it directly into the plot layer definitions,
and then the documentation can be done a bit more elegantly.