ttools: namespace multi-point layer types
Rename similar-sounding but different layer types, e.g. the Plane
and Sky variants of ellipse both used to be called "Ellipse", but
now they are "XYEllipse" and "SkyEllipse".
This means that they can have separately named, and separate sections
in the user document without needing complicated schemes to distinguish
them.