ttools: remove alternative DataGeoms from SkyPlotType
SkyPlotType used to have some alternative DataGeoms,
equatorial-galactic and galactic-equatorial for transforming
from data to view sky systems.  Since that transformation is
now handled by the viewsys and datasys parameters in SkyPlot2Task,
those are not necessary or useful.

SkyPlotType was the only one of the PlotTypes to use this facility
for multiple DataGeoms which is otherwise fiddly and in some places
nasty.  So I could remove it.  However, it might be desirable to
make use of it at some point, for instance to allow polar coordinates
in the Plane Plot, so don't do anything drastic for now.