vo: trim line responses from UWS services
For single-line responses like the output from the {jobs}/(job-id)/phase
UWS endpoint, add a switch to trim trailing whitespace from the result.
The GACS TAP service was returning phases like "COMPLETED\r\n" which
didn't match "COMPLETED" that I was looking for.  I don't know whether
this trailing whitespace is strictly legal, but I posted a message
on the GWS mailing list to find out, with no response, so give GACS
the benefit of the doubt for now.