ttools: plot2 tasks now include aux axis colour ramp
The AbstractPlot2Task interrogates the environment for global aux axis
config options and paints an Aux axis appropriately.

This is a bit messy - the aux config options are global as far as
the plot task goes, but per-layer for the plotters.
I've hacked around this for now by getting the AuxShader to report
its config options without using the global ones, but it means
that client code has to take special steps to put those in.
Maybe revisit the API or implementation at some stage.