ttools: reverse the polarity of legpos[1]
Invert the sense of the Y component of legend position so that
positive is up the page (like plot coordinates) not down the page
(like swing graphics coordinates).  It didn't used to matter when
the values were just used internally, but now it's exposed as the
legpos parameter we don't want to make users have to specify
legend Y positions the way they're not expecting.