ttools: extend capabilities of repeat filter
Added two features to the repeat filter: you can specify output row
sequence with the new optional flags -row|-table, and the repeat
count can be an expression rather than just a constant.

The first one is a sensible addition.  The second one is a bit niche,
and raises the question of whether all constants in STILTS should be
expressions.  I'd rather not, as it is more fiddly to implement,
and it means some usage errors don't get caught at argument parse time,
and possibly generate confusing error messages.  However, it's there
to support a use case from Baptiste Cecconi who wants to repeat rows
according to the length of an array-valued table parameter
(cmd="repeat size(param$Frequency)").