ttools: adjust classpath order for stilts invocation script
Put stilts.jar first on the list of places the stilts invocation script
looks for the ttools classes.  Previously topcat-full.jar was earlier,
but if a stilts release is separate from a topcat one that may be
a bad choice.  Not quite sure why I did it the other way round to start
with, but I think this is better.