ttools: fix taplint to accept application/xml as well as text/xml
In the UWS stage in some cases taplint used to require a content-type
of text/xml from some resources.  This looks like the sort of thing
I'd have put in because it was written in some standard somewhere,
but I can't find any trace of it.  So relax it to permit application/xml
as well - either is a reasonable way to indicate XML content.