ttools: treat unicodeChar like char for taplint type matching
taplint checks column type declarations in various places, and
reports a W_CTYP or E_CTYP in case of a mismatch.  This is done
according to the table in TAP v1.0 sec 2.5.  That table mentions
datatype "char" but not "unicodeChar", which is a bit unfair,
since it gives data providers no way to declare that they will
be emitting unicodeChar in VOTable output.

So this fix (lobbied for, entirely reasonably, by Markus) allows
unicodeChar content wherever char content would otherwise be permitted.
This is contrary to the letter, but in line with the spirit, of
the TAP standard.