ttools: taplint allows anyURI in dataModel ivo-id
There is an Erratum (#1) currently going through the system to redefine
the ivo-id attribute of TAPRegExt's dataModel element from
vr:identifierURI to xs:anyURI (so it can have a fragment identifier).

This commit reflects that erratum in taplint.

First, we need to adjust the TAPRegExt xml schema document,
changing the declared type of dataModel/@ivo-id,
so that the capabilities validation stage does not emit an error.
I've used a new TAPRegExt xsd document here labelled -Erratum, rather
than doing violence to the existing one; this version is not official,
I've doctored the existing one.

Then I added additional logging in the Capabilities stage to warn
if the ivorn does not match the regex corresponding to the
vr:identifierURI type.