table: facilitate DALI-compatible JDBC timestamp output
Add a utility method (createTimestampValueHandler) and a canned static
member (DALI) to the TypeMappers class, to cope with the case when
you want to serialize Timestamp columns from JDBC, and you want to
do it with a 'T' character separating the date and time parts of
the ISO-8601 string.  By default ("JDBC escape format") a space
character is used, but this is disallowed by the DALI standard,
hence DAL-family protocols such as TAP; they require a 'T' separator.