ttools: modify Coverage user functions
Add new function nearMoc, required to find out if a position is
within a given radius of a MOC, often more useful than whether
it strictly falls within it.  Also change the way that MOC names
are interpreted, so you can use a VizieR table ID as an alternative
to MOC FITS file URL/filename.