topcat: add doc links to history section
Up till now I have avoided using <ref> links in the history section
of SUN/253 on the grounds that the links may go out of date but the
history should be static.  This isn't a good enough argument to
justify it - the links are useful for people browsing what's changed.
So from now on I'm going to link up the history sections for SUN/253
and SUN/256 (but not backdating it).

In this commit add links for tapskymatch and GBIN input.