table: add alternative OnceRowPipe implementation
New class OnceRowPipe2.  This should provide identical behaviour to
OnceRowPipe, but it's implemented using more modern classes.
I wrote it to test something out.  It is not currently used
in any implementation.

For now this class is hidden in the API (it's package-private),
but I'm keeping it here in case OnceRowPipe causes trouble in the future
and I want something to compare it with.