ttools: adjust axis pan/zoom to avoid huge/tiny values
The Axis pan/zoom methods will now operate as no-ops in the case that
the output ranges would hit the limits of the double type,
i.e. if the exponent or mantissa runs out of bits.  This avoids
some problems to do with attempting to zoom in or out a ridiculous amount.