fits: fix PCOUNT bug in fits-var writer
In the VariableFitsTableSerializer, which uses P and Q output descriptors
to write variable-length arrays into the FITS HDU heap area,
I was writing the wrong value for the PCOUNT header card.
It was using the size of the heap, but it must include also the size
of any gap between the end of the data area and the start of the heap.
Fixed.