fits: add BasicInput framework for primitive stream input
Add the interface BasicInput along with implementations and factory
classes, providing an alternative mechanism for reading primitive
values from a byte sequence.

The previous mechanism used the java.io.DataInput class, whose usage
was inherited from the nom.tam.util classes.  Although nom.tam.*
is still in some use in this package, for bulk data access custom
classes are now used.  Abstracting required input methods to a
custom interface makes it clearer what methods are actually used
by the FITS table reader classes, so I can be sure these are
implemented optimally (the performance of these classes are core to
the performance of STIL on large tables).

The new mapped I/O implementations added here are, unlike the
previous ones, capable of relinquishing MappedByteBuffers during
the lifetime of a reader.  It looks like this is necessary on
some platforms to prevent system lockup from overmapping of memory.

This commit does not yet incorporate use of this new infrastructure
into the FITS reader classes.