fits: rewrite binary table input to use BasicInput classes
BintableStarTable is rewritten to use the BasicInput interface
for primitive data input rather than the ArrayDataInput abstraction.
This provides more control over release of resources.

The implementation has also changed to make more careful use of
data accessors - there is one thread-unsafe data reader for random
access (if provided), and one for each RowSequence, rather than
having a single MappedByteBuffer (or group of them) for random access
that needs synchronization.  That should give (and tests show that it does
give) improved performance, as well as a more obviously thread-safe
implementation.

This also allows less baroque use/re-use of input streams in the
FitsTableBuilder class, when pulling table data out of the input stream.

Quite a bit of reworking was required.