fits: rewrite colfits table input to use BasicInput classes
ColFitsStarTable is reworked to use the BasicInput interface for
primitive data input rather than MappedFile/MappedMultiFile.
This provides more control over release of resources.
It is more flexible (now it copes with stream input as well as
uncompressed files on disk, though using it like that is not necessarily
a good idea).

The implementation has also changed to make more careful use of
data accessors - there is one thread-unsafe data reader for random
access (if provided), and one for each RowSequence, rather than
having a single MappedByteBuffer (or group of them) for random access
that needs synchronization.  That should give (and tests show that it does
give) improved performance, as well as a more obviously thread-safe
implementation.

Quite a bit of reworking was required.