topcat: modify MacOS stub for oracle java
A user reports that attempting to use the topcat-full.dmg file
on MacOS Yosmite (OSX 10.10) complains:

   To open "TOPCAT.app" you need to install the legacy Java SE 6 runtime.

Some research suggests that this is because the way Apple java (java 6)
was installed differs from the way that Oracle java (java 7+) is done.
The official advice is to use Oracle's AppBuilder package to build the
app file, instead of JarBundler which is what I'm using.  I haven't
tried this (it claims to require ant 1.8.2, which we're not using),
but presumably it would work for only Oracle and not Apple javas,
which might disadvantage some people.  Relevant documents:

   https://java.net/projects/appbundler/
   http://docs.oracle.com/javase/7/docs/technotes/guides/jweb/packagingAppsForMac.html
   http://stackoverflow.com/questions/14806709/application-is-using-java-6-from-apple-instead-of-java-7-from-oracle-on-mac-os-x

Instead I'm trying out here UniversalJavaApplicationStub, which supplies
an alternative stub file (it's a shell script) that tries to work
with either form of installation:

   https://github.com/tofi86/universalJavaApplicationStub/