topcat: provide histogram export actions from histogram window
Actions are now provided in the plot2 Histogram window Export menu
to export the accumulated histogram bin data either to a new table file
or as a new table loaded instantly into the topcat application.

This reinstates functionality which was present in the old-style
histogram window.  It wasn't possible before now, because there
was previously no mechanism to pass the histogram bin data back
from the plotted layers to the GUI.  The new plotting ReportMap
framework now provides this.