topcat: add framework for explicit plot geometry setting
Add hooks to allow a PlotPosition object to be supplied from the GUI.
This can provide explicit settings for the size in pixels of the
generated plot graphic, as well as the borders between the exterior
of the graphic and the plot data region.  Any or all of these items
may be omitted, in which case defaults are calculated and used.

At this commit, no corresponding GUI has been implemented.