fits: add startable.unmap system property
The new startable.unmap system property should not be needed, but it's
added as an off-switch in case one is required for the MappedByteBuffer
unmapping machinery in the FITS package, related to the reworking
of the mapped file access.  Since the unmapping is dangerous
(could in principle cause JVM crashes if I haven't done it right),
it's a good idea to have a backup plan.