topcat: hack universalJavaApplicationStub to work?
The previous commit doesn't seem to help.  A comment from another
UniversalJavaApplicationStub user here:

   https://github.com/tofi86/universalJavaApplicationStub/issues/9

suggests that OSX is doing some magic with the Java part of the
Info.plist file that stops this approach working.

This commit (currently untested) changes the name of the relevant
section in the Info.plist file from Java to JavaUJAS (in accordance
with the suggestion above).  However, looking at the stub script,
this is really pretty pointless - if I want the thing to work with
both legacy (Apple) java 6 and current (Oracle) java 7+, I could
just write my own stub script which doesn't pull java config out
of the Info.plist file.

But it might be better to use Oracle's AppBuilder, or just to hardwire
the relevant files for the TOPCAT.app directory.  Looking at what
JarBundler does, which is probably not very different from AppBuilder,
it's not exactly magic.