ttools: fix javadoc syntax errors reported by java 8
Java 8 is stricter and more verbose about syntax errors in javadocs
than previous JDKs, so builds were failing.  Used the java 8 javadoc
error report to catch and fix various errors in ttools javadocs,
mostly using non-HTML syntax in javadocs
(e.g. unescaped > and & characters).

This fix is particularly necessary for the TTOOLS package when building
under java 8, since the TTOOLS build file has failonerror set true
for various javadoc invocations.  I *think* that's because doclets
are used as part of the automatic process that generates user
documentation ... but maybe it's just that it seemed like a good
idea to turn it on one day.