fits: add nasty hack around LDAC FITS format idiosyncracies
This is about the LDAC FITS variant:
http://marvinweb.astro.uni-bonn.de/data_products/THELIWWW/LDAC/LDAC_concepts.html

The first (LDAC_IMHEAD) table extension contains a single-cell table
where the cell's content is an array of FITS headers.
In auto-format-detection mode this confuses STIL into thinking
that the file is in colfits format, which in turn means that
the StarTableFactory.makeStarTables method ignores the second
(LDAC_OBJECTS) table extension, since colfits is not
multi-table-capable.

Add a hack to spot LDAC_IMHEAD HDUs in colfits and reject them as
colfits.