votable: improve Decoder behaviour for unknown datatypes
When a VOTable FIELD is encountered with an unknown (i.e. illegal)
value of the datatype attribute, an UnknownDecoder is used
(though the VOTable parser would be well within its rights to
abort the parse in that case).

Previously, this UnknownDecoder has a content class of String[].
This caused problems when serializing to FITS, since FITS can't
serialize arrays of uknown length - in general serializing arrays
is problematic.  The FITS serializer ended up throwing a
NullPointerException when attempting to serialize values from that
column, which was itself an error I haven't fixed.

Now, UnknownDecoder has a content class of (scalar) String.
This is much more sensible and easier to implement, and it
avoids (any?) problems with serialization.
I don't know why I didn't do it like that in the first place,
but from the history, it looks like it might be an editing
error in 2003(!).  Surprised it hasn't shown up since then.