ttools: rationalise and refactor RampKeySet
Make RampKeySet more flexible, and use it for the density shading
in ShapeMode as well as for Aux and Spectro shading.
This eliminates some special case/duplicated code in density shading,
and makes it easier to add extra shader functionality in a consistent
way for different functions.

RampKeySet.Ramp interface now just aggregates a Shader and a Scaling,
which is all it was really doing before.