topcat: add stilts-friendly tooltips to plot2 config labels
The JLabels attached to configuration specifiers and coordinate
selectors used in the plot2 windows now sport tooltips of
the form name=value, which can be used directly in the corresponding
STILTS plotting command.

The tooltips used to be on the the config (but not coord) specifiers
themselves, but this didn't work very well because in some cases
(specifiers composed of more than one component) they weren't
visible, so this has been withdrawn.