topcat: make plot2 config tooltips instant and persistent
Configure the stilts-friendly plot2 tooltips on config and coord
labels so that they pop up instantly and stay as long as the mouse
is on them, rather than delaying before they are seen and then
disappearing after a while.

I think this is a good idea, especially so that people notice this
information is available from the labels.  However, it might prove
to be annoying, in which case it could be harmlessly withdrawn
in future.