vo: use schema-level TAP metadata
Up till now, the TAP_SCHEMA interrogation and the parsing of the
/tables endpoint for TAP services has ignored schema-level metadata,
treating the metadata as a list of ungrouped table elements.

This change adds schema-awareness to the TAP infrastructure.

Some additional schema-related checks are added to the taplint
metadata validation.

This change makes possible changes to the user interfaces of other
downstream applications, but those changes have not yet been made.