ttools: allow NaN coordinates from Surface.dataToGraphics result
Clarify the contract and modify the PlaneSurface and CubeSurface
implementations of the Surface interface, so that input NaN data
coordinates can result in output NaN graphics coordinates *if*
visibleOnly is false.

Previously in some cases NaN outputs (perhaps resulting from NaN
inputs) resulted in summary failure (false return value).

That was problematic, since in some cases ("partial" coordinates,
used for things like histograms and spectrograms) the inputs
are supposed to be NaN in one coordinate and you still want a
non-failure result.  This was leading to failure of some things
that relied on partial coordinate processing in histogram/spectrogram
layers (Histogram/Time plot windows):

   - the "New subset from visible" button did nothing
   - the point count was always zero

That was a regression bug dating from commit c51d80215, so present
in v4.2-2 but not v4.2-1.

Following the fix, it's necessary to perform some more checking on
result graphics coordinates at the various places where
dataToGraphics (etc) is invoked with visibleOnly set false.