topcat: tweak MacOS application stub
Fix a bug in TopcatUJAStub: it was failing to change directory to the
user's home directory because of a shell syntax error in the
original universalJavaApplicationStub script.

This relates to the following issue/fix:
   https://github.com/tofi86/universalJavaApplicationStub/issues/11
   https://github.com/tofi86/universalJavaApplicationStub/commit/72373246c7e

Also avoid requesting a particular JVM version, since there seem to
be potential problems with that in UJAStub (probably OK but may be
safer to avoid it).  Since we only require >=v1.5, any available
MacOS version will be fine.