ttools: limit auto-range zoom for sky plot SinProjection
When auto-ranging a sky plot using the Sin projection with a set of
points separated by no more than a few milliarcseconds
(most often this was seen when plotting a single sky position),
the auto-ranging tried to zoom too far in and caused the
grid generation to fail with text to stderr.

Introduce a maximum zoom to apply when auto-ranging so that this
doesn't happen.