ttools: ConfigException is now checked, not unchecked
Reparent ConfigException to inherit from Exception not RuntimeException.
It's really necessary to have it as a checked exception to make it
clear when you have to handle it.  It was only unchecked before because
I was in a tearing hurry when I put did it in the first place.

Added some code in topcat to behave sensibly when such an exception
is caught, but this is not much tested so far; currently these are
not often thrown.