topcat: configure KDE smoothing/kernel centrally for Histogram
The KDE smoothing and kernel selection is now configured in the
Bars control of the Histogram window rather than per-layer.
The Bars control is now divided into Histogram, KDE and General
sections.

Also adjust the resizing code to take account of changes in KDE
layers as well as Histogram layers.