ttools: histogram-like plots can report actual bin widths
The bin size for a histogram and smoothing width for a KDE are
specified with a BinSizer config item.  In the GUI this is usually
done with a slider, which selects the relative widths of bins
in terms of the plot width.  The actual width in data coordinates
was not explicitly displayed to the user, because it's not known
at selection time (only when the plot is done).

Add machinery to the BinSizer config specifier so that the
relevant Plotter types can report back the actual widths in
data coordinates for display in the GUI.  This is useful visual
feedback for the user if he is using an unmarked slider for control.