fits: allow colfits files with missing TDIMn headers
If the colfits input handler encounters a FITS file which looks
OK except that it has missing TDIMn header cards, it will now assume
that those columns represent scalars, as long as the item counts
make sense.  Previously it threw a TableFormatException.

The previous behaviour was on the basis that if a file had been
written by the colfits output handler, it would have the TDIMn
headers written explicitly, and if it wasn't it wasn't likely
to be comprehensible.  However, at least some column-oriented
files written by other software, lacking TDIMn headers but
otherwise OK, seem to work fine, so it seems like a good idea
to lift the restriction.

This came up in relation to the following file from Galaxy Zoo:
http://galaxy-zoo-1.s3.amazonaws.com/schawinski_GZ_2010_catalogue.fits.gz
apparently "written by IDL".