util: SuperJar now deletes most META-INF entries for -oz option
When writing a jar-of-zips from SuperJar, any non-main attributes
in the manifest file, and any non-MANIFEST files in the META-INF
directory, are now ignored (not copied to output).

Previously the signatures were getting mangled and the jar files
were apparently unusable ("Invalid signature file digest for Manifest
main attributes").  I'm not quite sure why, but it may be because
the main-class attribute was getting rewritten.  In any case, as
used (for the stil-jars.zip distribution file) these signatures
are not worth the paper they're written on, and are best dropped.