topcat: histogram/KDE panels report visible layer count
You can now see how many histograms/KDEs are currently visible
in the control panels that configure the bin widths for the
histogram window.  This is mainly a good idea so that people can
notice there are no visible KDEs, which is why twiddling the knobs
on that panel doesn't do anything.