topcat: pass ReportMaps to specifiers in LayerControls
Add machinery to topcat plot windows to pass the post-plot reports
to the relevant specifiers that gave rise to the plots.

This does add what looks like additional run-time work;
in principle it might be a performance hit.  But although the
extra work looks complicated, I dont think it's actually going
to be noticeable in performance terms.