topcat: add visibility checkbox to subset-specific style control
In the Subset Styles panel (below the Global Style panel) of
FormStylePanel (the Form tab of multi-subset layer controls)
there is now a Visible checkbox next to the subset selector.

This isn't strictly necessary, but a user pointed out that otherwise
it's easy to change subset-specific style config and wonder why it's
not having any visible effect on the plot.  If there's an unchecked
visibility indicator there, it should be more obvious what's going on.