ttools: add KNN kernels to Kernel1dShape
An adaptive (variable-width) kernel based on the K-Nearest-Neighbours
scheme is added to the Kernel1dShape interface.

This looked like it was going to be a great idea, but having tried
it out, I'm not sure it's useful, since it wipes out the noise
in short bins, leaving the tall bins untouched.  At present this
new functionality is not accessible anywhere in the UI.