vo: make TAP metadata reading pluggable
The code to read TAP schema metadata is now abstracted to a new
interface TapMetaReader.  Currently the only implementation is,
as before, by reading the /tables endpoint, but this allows more
flexible schemes (e.g. TAP_SCHEMA table queries).

The asynchronous code that queries for schema metadata is moved
from TapQueryPanel to TableSetPanel for tidiness, since that's
where (as yet unwritten) code for further queries of table
details will need to be done.