vo: allow asynchronous table metadata updates
TableSetPanel will now spot missing column and foreign-key metadata
for displayed tables and acquire it asynchronously, displaying the
relevant items when ready.

This is not required for the TableSetTapMetaReader, but TapMetaReader
implementations which do not fill in all table objects will need it.

A couple of things remain to be done.
Really the TableSetPanel should have more logic to prevent it
dispatching new requests for a metadata item that has been requested
but not yet returned.
Also the AdqlValidator class needs to be updated to cope with
initially-columnless table descriptions.