ttools: add new semi_steps histogram bar style
This is like the recently added semi_filled style, but without bar
boundary lines going all the way to the axis.
The rendering to vector formats like EPS and PDF is currently not
very good, because of apparent imprecision in graphic placement,
leading to regions where the transparent colour is overplotted
twice so looks darker.  To fix this, I'd have to change the histogram
implementation to plot all the bars at once rather than a bar at
a time.