ttools: allow slider text field to reflect slider state
Extend the options for SliderSpecifier so that if the text entry
box is displayed, you can choose to have it updated with the
current value of the slider, if the slider is active.
This gives useful user feedback (though it does compress the size
of the slider a bit in narrow components).

The following sliders are updated so that they now provide this
feedback:
   RampKeySet quantisation
   StyleKeys.SCALE and SCALE_PIX shape scale factor keys
   KNN KDE plotter K selection key