ttools: fix AREA normalisation mode to work with log X axis
I previously hobbled AREA when the X axis is logarithmic,
since it can't normalise to data units.
But there's no reason it can't work with log data units.
Fix it so it does, and make it the default normalisation option
(apart from NONE), since it's probably the most useful one.