vo: use Doug Lea's LinkedBlockingStack class
Download LinkedBlockingStack from here:

   http://gee.cs.oswego.edu/dl/code/LinkedBlockingStack.java

This is a utility class that implements BlockingQueue with LIFO
ordering rather than FIFO.  This commit is the class as downloaded,
so it doesn't even have a suitable namespace.

Note, at J2SE 1.6 there exists a class
java.util.concurrent.LinkedBlockingDeque which doesn't quite do
the same thing, but could be adapted for the same purpose.
It may be sensible to remove this class in favour of that if/when
the starjava codebase targets 1.6.