util: allow log level configuration for resource read failure
You can now configure at what level a log level is written in case
of failure of the IOUtils.getResourceContents method.

Use this to fix the version reporting (-version command-line flag)
on topcat and stilts so that it doesn't issue a warning message
if the git revision string is absent (since it sometimes is absent
under normal circumstances).