vo: sort schemas and tables acquired from TAP_SCHEMA
When reading TAP metadata from TAP_SCHEMA, sort the list of schemas,
and the list of tables within each schema, alphabetically by name.

Currently, this is not done when reading metadata from the /tables
endpoint, since the sequence of tables in that case may have been
arranged by the service in a particular order (unlike database
tables, which are in principle unordered).  But this does result
in an inconsistency between the display of the two ways of acquiring
service metadata, so possibly I ought to do the same for /tables
as well.  Hmm.